refactor: add app bars in package and refactor custom app bar
