+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/design_system/component/appbar/ActionIconButton.kt
@@ -1,22 +1,21 @@
-package com.cairosquad.evolvefit.design_system.component
+package com.cairosquad.evolvefit.design_system.component.appbar
 
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
-import org.jetbrains.compose.resources.painterResource
-import org.jetbrains.compose.resources.DrawableResource
+import androidx.compose.ui.graphics.painter.Painter
 
 @Composable
 fun ActionIconButton(
-    icon: DrawableResource,
+    icon: Painter,
     contentDescription: String?,
-    tint: Color = Color.Unspecified,
-    onClick: () -> Unit
+    onClick: () -> Unit,
+    tint: Color = Color.Unspecified
 ) {
     IconButton(onClick = onClick) {
         Icon(
-            painter = painterResource(icon),
+            painter = icon,
             contentDescription = contentDescription,
             tint = tint
         )
