+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/design_system/component/appbar/LanguageSelectorBar.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.evolvefit.design_system.component
+package com.cairosquad.evolvefit.design_system.component.appbar
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.tween
